Go back to using the original, self-pipe based implementation of interruptible
For all Rubies, and not just Ruby < 3.2.
This is looking like the most likely culprit of a memory leak pointed
out for multiple people on https://github.com/rails/solid_queue/issues/330
