+++ lib/solid_queue/processes/base.rb
@@ -4,7 +4,7 @@ module SolidQueue
   module Processes
     class Base
       include Callbacks # Defines callbacks needed by other concerns
-      include AppExecutor, Registrable, Procline
+      include AppExecutor, Registrable, Interruptible, Procline
 
       attr_reader :name
